+++ nvim/lua/plugins.lua
@@ -1,4 +1,4 @@
-return require('packer').startup(function()
+return require('packer').startup(function(use)
   use 'wbthomason/packer.nvim'
 
   use 'tpope/vim-commentary'
@@ -20,7 +20,8 @@ return require('packer').startup(function()
     'lewis6991/gitsigns.nvim',
     requires = {
       'nvim-lua/plenary.nvim'
-    }
+    },
+    commit = '5655cfa4c0316943c7da641f813e79a91bc0d87e'
   }
   use 'Raimondi/delimitMate'
   use 'jeetsukumaran/vim-indentwise'
@@ -30,6 +31,25 @@ return require('packer').startup(function()
     'neoclide/coc.nvim',
     branch = 'release'
   }
+
+  use 'michaeljsmith/vim-indent-object'
+
+  use 'mfussenegger/nvim-dap'
+	use 'nvim-telescope/telescope-dap.nvim'
+	use 'theHamsta/nvim-dap-virtual-text'
+  use 'Pocco81/DAPInstall.nvim'
+
+  use 'szw/vim-maximizer'
+  use 'AndrewRadev/linediff.vim'
+
+  use {
+    "folke/zen-mode.nvim",
+    config = function()
+      require("zen-mode").setup {}
+    end
+  }
+
+  use "wellle/context.vim"
 
 end)