+++ nvim/plugin/packer_compiled.lua
@@ -69,9 +69,17 @@ end
 time([[try_loadstring definition]], false)
 time([[Defining packer_plugins]], true)
 _G.packer_plugins = {
+  ["DAPInstall.nvim"] = {
+    loaded = true,
+    path = "/Users/alexignez/.local/share/nvim/site/pack/packer/start/DAPInstall.nvim"
+  },
   ["coc.nvim"] = {
     loaded = true,
     path = "/Users/alexignez/.local/share/nvim/site/pack/packer/start/coc.nvim"
+  },
+  ["context.vim"] = {
+    loaded = true,
+    path = "/Users/alexignez/.local/share/nvim/site/pack/packer/start/context.vim"
   },
   ["defx.nvim"] = {
     loaded = true,
@@ -85,9 +93,21 @@ _G.packer_plugins = {
     loaded = true,
     path = "/Users/alexignez/.local/share/nvim/site/pack/packer/start/gitsigns.nvim"
   },
+  ["linediff.vim"] = {
+    loaded = true,
+    path = "/Users/alexignez/.local/share/nvim/site/pack/packer/start/linediff.vim"
+  },
   ["nvim-base16"] = {
     loaded = true,
     path = "/Users/alexignez/.local/share/nvim/site/pack/packer/start/nvim-base16"
+  },
+  ["nvim-dap"] = {
+    loaded = true,
+    path = "/Users/alexignez/.local/share/nvim/site/pack/packer/start/nvim-dap"
+  },
+  ["nvim-dap-virtual-text"] = {
+    loaded = true,
+    path = "/Users/alexignez/.local/share/nvim/site/pack/packer/start/nvim-dap-virtual-text"
   },
   ["nvim-treesitter"] = {
     loaded = true,
@@ -104,6 +124,10 @@ _G.packer_plugins = {
   ["popup.nvim"] = {
     loaded = true,
     path = "/Users/alexignez/.local/share/nvim/site/pack/packer/start/popup.nvim"
+  },
+  ["telescope-dap.nvim"] = {
+    loaded = true,
+    path = "/Users/alexignez/.local/share/nvim/site/pack/packer/start/telescope-dap.nvim"
   },
   ["telescope.nvim"] = {
     loaded = true,
@@ -125,9 +149,17 @@ _G.packer_plugins = {
     loaded = true,
     path = "/Users/alexignez/.local/share/nvim/site/pack/packer/start/vim-fugitive"
   },
+  ["vim-indent-object"] = {
+    loaded = true,
+    path = "/Users/alexignez/.local/share/nvim/site/pack/packer/start/vim-indent-object"
+  },
   ["vim-indentwise"] = {
     loaded = true,
     path = "/Users/alexignez/.local/share/nvim/site/pack/packer/start/vim-indentwise"
+  },
+  ["vim-maximizer"] = {
+    loaded = true,
+    path = "/Users/alexignez/.local/share/nvim/site/pack/packer/start/vim-maximizer"
   },
   ["vim-obsession"] = {
     loaded = true,
@@ -140,10 +172,19 @@ _G.packer_plugins = {
   vimwiki = {
     loaded = true,
     path = "/Users/alexignez/.local/share/nvim/site/pack/packer/start/vimwiki"
+  },
+  ["zen-mode.nvim"] = {
+    config = { "\27LJ\2\n:\0\0\3\0\3\0\a6\0\0\0'\2\1\0B\0\2\0029\0\2\0004\2\0\0B\0\2\1K\0\1\0\nsetup\rzen-mode\frequire\0" },
+    loaded = true,
+    path = "/Users/alexignez/.local/share/nvim/site/pack/packer/start/zen-mode.nvim"
   }
 }
 
 time([[Defining packer_plugins]], false)
+-- Config for: zen-mode.nvim
+time([[Config for zen-mode.nvim]], true)
+try_loadstring("\27LJ\2\n:\0\0\3\0\3\0\a6\0\0\0'\2\1\0B\0\2\0029\0\2\0004\2\0\0B\0\2\1K\0\1\0\nsetup\rzen-mode\frequire\0", "config", "zen-mode.nvim")
+time([[Config for zen-mode.nvim]], false)
 if should_profile then save_profiles() end
 
 end)