+++ nvim/ultisnips/extensions/coc-sumneko-lua-data/sumneko-lua-ls/extension/server/platform.lua
@@ -11,8 +11,12 @@ end
 local exePath = findExePath()
 local exeDir  = exePath:match('(.+)[/\\][%w_.-]+$')
 local dll     = package.cpath:match '[/\\]%?%.([a-z]+)'
-package.cpath = ('%s/?.%s'):format(exeDir, dll)
-local ok, err = package.loadlib(exeDir..'/bee.'..dll, 'luaopen_bee_platform')
+package.cpath = ('%s/?.%s;%s'):format(exeDir, dll, package.cpath)
+local bee     = package.searchpath('bee', package.cpath)
+if not bee then
+    error('Can not find bee.dll? cpath = ' .. tostring(package.cpath))
+end
+local ok, err = package.loadlib(bee, 'luaopen_bee_platform')
 if not ok then
     error(([[It doesn't seem to support your OS, please build it in your OS, see https://github.com/sumneko/vscode-lua/wiki/Build
 errorMsg: %s