+++ nvim/ultisnips/extensions/coc-sumneko-lua-data/sumneko-lua-ls/extension/server/script/utility.lua
@@ -21,6 +21,7 @@ local mathHuge     = math.huge
 local inf          = 1 / 0
 local nan          = 0 / 0
 local utf8         = utf8
+local error        = error
 
 _ENV = nil
 
@@ -642,9 +643,45 @@ function m.expandPath(path)
             home = getenv 'USERPROFILE' or (getenv 'HOMEDRIVE' .. getenv 'HOMEPATH')
         end
         return home .. path:sub(2)
-    else
+    elseif path:sub(1, 1) == '$' then
+        path = path:gsub('%$([%w_]+)', getenv)
         return path
     end
+    return path
+end
+
+function m.arrayToHash(l)
+    local t = {}
+    for i = 1, #l do
+        t[l[i]] = true
+    end
+    return t
+end
+
+function m.switch()
+    local map = {}
+    local cachedCases = {}
+    local obj = {
+        case = function (self, name)
+            cachedCases[#cachedCases+1] = name
+            return self
+        end,
+        call = function (self, callback)
+            for i = 1, #cachedCases do
+                local name = cachedCases[i]
+                cachedCases[i] = nil
+                if map[name] then
+                    error('Repeated fields:' .. tostring(name))
+                end
+                map[name] = callback
+            end
+            return self
+        end,
+        getMap = function (self)
+            return map
+        end
+    }
+    return obj
 end
 
 return m