+++ nvim/ultisnips/extensions/coc-sumneko-lua-data/sumneko-lua-ls/extension/server/script/vm/getGlobals.lua
@@ -1,10 +1,17 @@
 local collector = require 'core.collector'
----@type vm
+local guide     = require 'parser.guide'
+---@diagnostic disable-next-line
+---@class vm
 local vm        = require 'vm.vm'
 local noder     = require 'core.noder'
 
 function vm.hasGlobalSets(name)
-    local id = ('def:g:%q'):format(name)
+    local id
+    if type(name) == 'string' then
+        id = ('def:g:%s%s'):format(noder.STRING_CHAR, name)
+    else
+        id = ('def:g:%s'):format(noder.STRING_CHAR, name)
+    end
     return collector.has(id)
 end
 
@@ -15,15 +22,32 @@ function vm.getGlobalSets(name)
     end
     local results = {}
     cache[name] = results
-    local id
     if name == '*' then
-        id = 'def:g:'
+        for noders in collector.each('def:g:') do
+            for id in noder.eachID(noders) do
+                if  id:sub(1, 2) == 'g:'
+                and not id:find(noder.SPLIT_CHAR) then
+                    for source in noder.eachSource(noders, id) do
+                        if guide.isSet(source) then
+                            results[#results+1] = source
+                        end
+                    end
+                end
+            end
+        end
     else
-        id = ('def:g:%q'):format(name)
-    end
-    for node in collector.each(id) do
-        for source in noder.eachSource(node) do
-            results[#results+1] = source
+        local id
+        if type(name) == 'string' then
+            id = ('g:%s%s'):format(noder.STRING_CHAR, name)
+        else
+            id = ('g:%s'):format(noder.STRING_CHAR, name)
+        end
+        for noders in collector.each('def:' .. id) do
+            for source in noder.eachSource(noders, id) do
+                if guide.isSet(source) then
+                    results[#results+1] = source
+                end
+            end
         end
     end
     return results